Updates picture load
Adds a div, pic-wrapper, around the image so that the space is already
taken up. I was noticing that when loading the site if the picture
doesn't load immediately on the about me page, the styles would jump
around. Adding some permanent blank space in the form of a div should
resolve this issue.